implemented to load image in DetailStatusView
