+++ app/src/main/java/com/freshdigitable/udonroad/DetailStatusView.java
@@ -29,17 +29,11 @@ import android.widget.TextView;
 import java.util.List;
 
 import twitter4j.Status;
-import twitter4j.User;
 
 /**
  * Created by akihit on 2016/08/18.
  */
-public class DetailStatusView extends StatusViewBase {
-  protected TextView rtUser;
-  protected ImageView rtUserIcon;
-  protected LinearLayout rtUserContainer;
-  private QuotedStatusView quotedStatus;
-
+public class DetailStatusView extends FullStatusView {
   public DetailStatusView(Context context) {
     this(context, null);
   }
@@ -74,41 +68,6 @@ public class DetailStatusView extends StatusViewBase {
   }
 
   @Override
-  public void bindStatus(final Status status) {
-    super.bindStatus(status);
-    if (status.isRetweet()) {
-      bindRtUser(status.getUser());
-    }
-
-    final Status quotedBindingStatus = getBindingStatus(status).getQuotedStatus();
-    if (quotedBindingStatus != null) {
-      quotedStatus.bindStatus(quotedBindingStatus);
-      quotedStatus.setVisibility(VISIBLE);
-    }
-  }
-
-  private void bindRtUser(User user) {
-    setRetweetedUserVisibility(VISIBLE);
-    final String formattedRtUser = formatString(R.string.tweet_retweeting_user,
-        user.getScreenName());
-    rtUser.setText(formattedRtUser);
-  }
-
-  private void setRetweetedUserVisibility(int visibility) {
-    rtUserContainer.setVisibility(visibility);
-  }
-
-  @Override
-  public void reset() {
-    super.reset();
-    setRetweetedUserVisibility(GONE);
-    rtUserIcon.setImageDrawable(null);
-    quotedStatus.setBackgroundResource(R.drawable.s_quoted_frame);
-    quotedStatus.setVisibility(GONE);
-    quotedStatus.reset();
-  }
-
-  @Override
   protected CharSequence parseText(Status status) {
     final Status bindingStatus = getBindingStatus(status);
     String text = bindingStatus.getText();
@@ -127,14 +86,6 @@ public class DetailStatusView extends StatusViewBase {
         ssb.replace(si.start, si.end, si.displayingText);
       }
     }
-  }
-
-  public ImageView getRtUserIcon() {
-    return rtUserIcon;
-  }
-
-  public QuotedStatusView getQuotedStatusView() {
-    return quotedStatus;
   }
 
   @Override
